Move render and apply to pkg
