+++ controllers/poisonpillconfig_controller.go
@@ -21,8 +21,8 @@ import (
 	"fmt"
 	"github.com/go-logr/logr"
 	poisonpillv1alpha1 "github.com/medik8s/poison-pill/api/v1alpha1"
-	"github.com/medik8s/poison-pill/apply"
-	"github.com/medik8s/poison-pill/render"
+	"github.com/medik8s/poison-pill/pkg/apply"
+	"github.com/medik8s/poison-pill/pkg/render"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
