+++ pkg/render/testdata/machineconfig/files/configure-switchdev.sh.yaml
