+++ pkg/render/testdata/machineconfig/units/NetworkManager.service.yaml
