+++ pkg/render/testdata/machineconfig/units/switchdev-configuration.yaml
