+++ pkg/render/testdata/manifests/multiple.yaml
