+++ pkg/render/testdata/manifests/simple.json
