+++ pkg/render/testdata/manifests/simple.yaml
