+++ pkg/render/testdata/manifests/template.yaml
